requirements.txt
openmeteo-requests
requests-cache
retry-requests
numpy
pandas
pylint
python-dotenv
pytest